+++ geolib/__init__.py
@@ -2,7 +2,7 @@
 GEOLib Library
 """
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 
 from . import utils
 from .models import *